fix(dashboard): round credits to 2dp and monthly spend to whole numbers
Credits and monthly spend were leaking raw floating-point precision
(e.g. 984.9695, 1065.774399999...) into API responses. Round credits
to 2 decimal places across the dashboard, workspace, and recruitment
dashboard schemas, and make dashboard monthly_spent a whole number.

+++ app/schemas/dashboard.py
@@ -39,7 +39,7 @@ class DashboardRecentCallItem(BaseModel):
 class DashboardSummary(BaseModel):
     """All data needed to render the dashboard home view in a single response."""
 
-    monthly_spent: float
+    monthly_spent: int
     monthly_calls: int
     success_rate_percent: float | None
     credits: float
